Compute in f64, and fix the four fidelity bugs that exposed
Mismatches against the full catalogue: 113 -> 0.

Lengths, percentages and ratios were f32, so `h-1/3` emitted 33.333336%
against Tailwind's 33.33333333333333% and `-scale-105` emitted -104.99999%.
Sixty of the mismatches were that. The IR and the utility table are f64
now; nothing in either emitter needed changing.

Three more fell out once the noise was gone:

`-scale-0` emitted `scale: -0% -0%`. IEEE negation of zero is -0.0, which
Rust prints as `-0`; Tailwind writes `0`. The values behave identically in
CSS and the strings don't, so the `-` prefix now leaves zero unsigned.

`scale-110` emitted 110.00000000000001%. The IR held scale as a ratio, so
the Web lowering divided by 100 at parse and multiplied by 100 at emit --
a round trip through a value that isn't representable. It holds the
authored percentage now, which is exact for Web; React Native divides once,
where the same rounding is invisible because it takes a number, not a
string.

`transition-none` emitted a timing function and duration alongside the
property. Tailwind emits the property alone -- reasonably, since the point
is to turn transitions off.

Two of the reported mismatches turned out to be the harness, not Dowel. Its
`flex` shorthand expansion read the two-value form positionally, but the
spec resolves it by type: a number is the shrink factor, anything else is
the basis. `flex: 0 auto` is exactly `flex: 0 1 auto`, so Dowel was right.
And `shadow-none` is now a documented accepted difference -- Dowel emits
`none` where Tailwind clears its register and leaves a transparent ring
chain, which renders the same. Noted as *not* equivalent forever: once
`ring-*` exists, `shadow-none ring-2` would need Tailwind's form.

Verified: 135 Rust tests, 40 JS tests, curated report unchanged (Web
120/120 covered and matching, Native 104/120), login-demo builds.

+++ crates/dowel_ir/src/lib.rs
@@ -282,7 +282,7 @@ pub enum StyleProperty {
 
     // Visual
     BackgroundColor(Color),
-    Opacity(f32),
+    Opacity(f64),
     BorderColor(Color),
     // Per-side, for the same reason margin/padding are (see above):
     // `border-t-2` and `border-b-4` set disjoint sides and must compose.
@@ -316,7 +316,14 @@ pub enum StyleProperty {
     /// defined application order (translate, then rotate, then scale).
     Rotate(Angle),
     /// A ratio: `scale-95` is 0.95, not 95.
-    Scale(f32),
+    /// As a *percentage*, the way Tailwind writes it (`scale-110` -> 110).
+    ///
+    /// Held in the authored unit rather than as a ratio so the Web
+    /// lowering is exact: converting to a ratio at parse and back at emit
+    /// made `scale-110` come out as `110.00000000000001%`. React Native
+    /// wants the ratio and divides once, where the same rounding is
+    /// invisible -- it takes a number, not a string.
+    Scale(f64),
     TranslateX(Length),
     TranslateY(Length),
     /// Kept as the already-composed CSS value rather than a structured
@@ -378,7 +385,7 @@ pub enum FlexShorthand {
     Initial,
     Auto,
     None,
-    Grow(f32),
+    Grow(f64),
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -522,13 +529,13 @@ pub enum Radius {
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Length {
-    Px(f32),
+    Px(f64),
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Dimension {
     Length(Length),
-    Percent(f32),
+    Percent(f64),
     Auto,
     /// A percentage of the viewport (`h-screen` is `ViewportHeight(100.0)`).
     /// React Native has no viewport unit -- screen size there is a runtime
@@ -536,8 +543,8 @@ pub enum Dimension {
     /// can't hold -- so these are Web-only and the Native backend refuses
     /// them rather than freezing a launch-time size that would go stale on
     /// rotation.
-    ViewportWidth(f32),
-    ViewportHeight(f32),
+    ViewportWidth(f64),
+    ViewportHeight(f64),
 }
 
 impl Dimension {
@@ -566,13 +573,13 @@ pub enum TextAlign {
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Angle {
-    pub degrees: f32,
+    pub degrees: f64,
 }
 
 /// A length in `em` -- relative to the element's own font size, so it can't
 /// be resolved to pixels at compile time.
 #[derive(Debug, Clone, Copy, PartialEq)]
-pub struct Em(pub f32);
+pub struct Em(pub f64);
 
 /// CSS allows a line height to be an absolute length or a unitless
 /// multiplier of the font size. Tailwind uses both: `leading-6` is the
@@ -583,7 +590,7 @@ pub enum LineHeight {
     /// React Native has no unitless line height -- its `lineHeight` is an
     /// absolute number -- and the font size to multiply by isn't known at
     /// compile time, so this form is Web-only.
-    Ratio(f32),
+    Ratio(f64),
 }
 
 /// Tailwind's built-in animations. Named rather than stored as shorthand
